ARIES-975: semantic versioning tool serialisation id comparision is incorrect
git-svn-id: https://svn.apache.org/repos/asf/aries/trunk/versioning/versioning-checker@1415271 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/aries/versioning/utils/ClassDeclaration.java
@@ -30,7 +30,6 @@ import java.util.Map;
 import java.util.Set;
 
 import org.objectweb.asm.ClassReader;
-import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.Type;
 
 public class ClassDeclaration extends GenericDeclaration {
@@ -398,7 +397,7 @@ public class ClassDeclaration extends GenericDeclaration {
             if ((!getAllSupers().contains(SemanticVersioningUtils.ENUM_CLASS) && (!old.getAllSupers().contains(SemanticVersioningUtils.ENUM_CLASS)))) {
                 Long oldValue = getSerialVersionUID(old);
                 Long curValue = getSerialVersionUID(this);
-                if ((oldValue != curValue)) {
+                if ((oldValue.longValue() != curValue.longValue())) {
                     reasons.add("The serializable class is no longer back compatible as the value of SerialVersionUID has changed from " + oldValue + " to " + curValue + ".");
                 }
             }
